Add html_safe option to flash notice and alerts
In order to add links into flash messages we need to mark the message as
html_safe.

We initially had this set to true on all flash alerts but not flash
notices since none of our flash alerts had user inputted values so we
could feel safe doing this.

However this remaining safe relied on knowledge that flash alerts were
always marked as html_safe, which could be missed further in the future.

It also meant that we could not add html into flash notices as they were
never marked as html_safe, since they sometimes contained user input.

To make this a bit safer and more flexible we have added an opt-in key
to flash messages `html_safe` which can be marked as true in instances
where we need to render html in flash messages. This means that we have
to choose if we trust the content of the flash (i.e. it has no user
input).